requirements.txt
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=5.4
openai>=1.0.0
anthropic>=0.20.0
tqdm>=4.60.0
